Eliminate subxt
After #120 it is now possible to remove subxt.

* We create `client/src/error.rs` that defines the error enum that
  replaces `substrate_subxt::Error`. Note that we don’t need all the
  error cases.
* In `RemoteNode` we replace `subxt_client` with the underyling RPC
  clients.

Closes #109
