--- .gitmodules
@@ -1,3 +1,0 @@
-[submodule "subxt/vendor"]
-	path = subxt/vendor
-	url = https://github.com/paritytech/substrate-subxt.git
