+++ rustfmt.toml
@@ -1,5 +1,1 @@
 edition = "2018"
-ignore = [
-  "substrate",
-  "subxt",
-]
