+++ scripts/build-client-docs
@@ -7,7 +7,6 @@ set -euo pipefail
 cargo doc \
   --no-deps \
   -p radicle-registry-client \
-  -p substrate-subxt \
   -p substrate-primitives \
   -p sr-io \
   -p sr-primitives \
