+++ scripts/check-license-headers
@@ -1,6 +1,6 @@
 #!/usr/bin/env bash
 #
-# Iterate over all rust source files (except our `subxt` copy) and make sure the
+# Iterate over all rust source files and make sure the
 # files include the GPL3 license header.
 #
 # If run with the `--fix` option license headers are prepended if they are
@@ -13,10 +13,6 @@ license_header=$(cat ./LICENSE_HEADER)
 
 shopt -s globstar
 for file in */src/**/*.rs; do
-  if [[ "$file" = subxt/* ]]; then
-    continue
-  fi
-
   file_start=$(head -n 15 "$file")
 
   if [[ "$file_start" != "$license_header" ]]; then
