--- subxt/Cargo.toml
@@ -1,89 +1,0 @@
-[package]
-name = "substrate-subxt"
-version = "0.2.0"
-authors = ["Parity Technologies <admin@parity.io>"]
-edition = "2018"
-
-license = "GPL-3.0"
-readme = "README.md"
-
-description = "Submit extrinsics (transactions) to a substrate node via RPC"
-keywords = ["parity", "substrate", "blockchain"]
-
-include = ["/Cargo.toml", "src/**/*.rs", "/README.md", "/LICENSE"]
-
-[dependencies]
-derive_more = "0.14.0"
-futures = "0.1.28"
-jsonrpc-core-client = { version = "14.0", features = ["ws"] }
-log = "0.4"
-num-traits = { version = "0.2", default-features = false }
-parity-scale-codec = { version = "1.0", default-features = false, features = ["derive", "full"] }
-serde = { version = "1.0", features = ["derive"] }
-url = "1.7"
-
-[dependencies.runtime_metadata]
-package = "paint-metadata"
-git = "https://github.com/paritytech/substrate"
-rev = "dc4a8329b9c47790c25429a5df272b3648275c1b"
-
-[dependencies.runtime_primitives]
-package = "sr-primitives"
-git = "https://github.com/paritytech/substrate"
-rev = "dc4a8329b9c47790c25429a5df272b3648275c1b"
-
-[dependencies.runtime_support]
-package = "paint-support"
-git = "https://github.com/paritytech/substrate"
-rev = "dc4a8329b9c47790c25429a5df272b3648275c1b"
-
-[dependencies.sr-version]
-git = "https://github.com/paritytech/substrate"
-rev = "dc4a8329b9c47790c25429a5df272b3648275c1b"
-
-[dependencies.paint-balances]
-git = "https://github.com/paritytech/substrate"
-rev = "dc4a8329b9c47790c25429a5df272b3648275c1b"
-
-[dependencies.paint-contracts]
-git = "https://github.com/paritytech/substrate"
-rev = "dc4a8329b9c47790c25429a5df272b3648275c1b"
-
-[dependencies.paint-indices]
-git = "https://github.com/paritytech/substrate"
-rev = "dc4a8329b9c47790c25429a5df272b3648275c1b"
-
-[dependencies.paint-system]
-git = "https://github.com/paritytech/substrate"
-rev = "dc4a8329b9c47790c25429a5df272b3648275c1b"
-
-[dependencies.substrate-primitives]
-git = "https://github.com/paritytech/substrate"
-rev = "dc4a8329b9c47790c25429a5df272b3648275c1b"
-
-[dependencies.substrate-rpc-api]
-git = "https://github.com/paritytech/substrate"
-rev = "dc4a8329b9c47790c25429a5df272b3648275c1b"
-
-[dependencies.substrate-rpc-primitives]
-git = "https://github.com/paritytech/substrate"
-rev = "dc4a8329b9c47790c25429a5df272b3648275c1b"
-
-[dependencies.txpool]
-package = "substrate-transaction-graph"
-git = "https://github.com/paritytech/substrate"
-rev = "dc4a8329b9c47790c25429a5df272b3648275c1b"
-
-
-[dev-dependencies]
-env_logger = "0.6"
-tokio = "0.1"
-wabt = "0.9.0"
-
-[dev-dependencies.node-runtime]
-git = "https://github.com/paritytech/substrate"
-rev = "dc4a8329b9c47790c25429a5df272b3648275c1b"
-
-[dev-dependencies.substrate-keyring]
-git = "https://github.com/paritytech/substrate"
-rev = "dc4a8329b9c47790c25429a5df272b3648275c1b"
